grok_bin_oct_hex: Add two output flags
This commit adds two output flags returned from this function to the one
previously existing, so that the caller can be informed of the problems
found and take its own action, rather than having to deal with the
default action from this function.

This involves the behavior of two existing flags, whose being set
suppresses the warning if particular conditions exist in the input being
parsed.  Both flags are currently always cleared upon return.

One of those flags is non-public.  I changed it so that it isn't cleared
upon return if the condition it describes is found.

The other flag is public.  I thought that some existing code, though
unlikely,  might be relying on the flag being always cleared.  So I
added a completely new flag from a currently unused bit that, if set on
input, will be set on output if the condition is met; otherwise cleared.
The only code that could possibly be affected is that which sets this
unused bit, but expects it to be cleared after the return.  This is very
unlikely.